refactor : Menugroup 도메인 패키지 분리, Jpa 적용
