--- src/main/java/kitchenpos/application/MenuGroupService.java
@@ -1,30 +1,0 @@
-package kitchenpos.application;
-
-import kitchenpos.application.dto.MenuGroupCreateRequest;
-import kitchenpos.application.dto.MenuGroupResponse;
-import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.domain.MenuGroup;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-public class MenuGroupService {
-    private final MenuGroupDao menuGroupDao;
-
-    public MenuGroupService(final MenuGroupDao menuGroupDao) {
-        this.menuGroupDao = menuGroupDao;
-    }
-
-    @Transactional
-    public MenuGroupResponse create(final MenuGroupCreateRequest request) {
-        MenuGroup menuGroup = new MenuGroup(request.getName());
-        MenuGroup savedMenuGroup = menuGroupDao.save(menuGroup);
-        return MenuGroupResponse.of(savedMenuGroup);
-    }
-
-    public List<MenuGroupResponse> list() {
-        return MenuGroupResponse.list(menuGroupDao.findAll());
-    }
-}
