+++ src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,0 +1,30 @@
+package kitchenpos.menugroup.application;
+
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.dto.MenuGroupCreateRequest;
+import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.menugroup.repository.MenuGroupRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroupResponse create(final MenuGroupCreateRequest request) {
+        MenuGroup menuGroup = new MenuGroup(request.getName());
+        MenuGroup savedMenuGroup = menuGroupRepository.save(menuGroup);
+        return MenuGroupResponse.of(savedMenuGroup);
+    }
+
+    public List<MenuGroupResponse> list() {
+        return MenuGroupResponse.list(menuGroupRepository.findAll());
+    }
+}
