+++ src/main/java/kitchenpos/menugroup/domain/MenuGroup.java
@@ -1,8 +1,16 @@
-package kitchenpos.domain;
+package kitchenpos.menugroup.domain;
 
 import org.springframework.util.StringUtils;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
 public class MenuGroup {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
