+++ src/test/java/kitchenpos/menugroup/application/MenuGroupServiceTest.java
@@ -1,9 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.menugroup.application;
 
-import kitchenpos.application.dto.MenuGroupCreateRequest;
-import kitchenpos.application.dto.MenuGroupResponse;
-import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.dto.MenuGroupCreateRequest;
+import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.menugroup.repository.MenuGroupRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -19,7 +19,7 @@ class MenuGroupServiceTest {
     private MenuGroupService menuGroupService;
 
     @Autowired
-    private MenuGroupDao menuGroupDao;
+    private MenuGroupRepository menuGroupRepository;
 
     @DisplayName("메뉴그룹을 생성한다.")
     @Test
@@ -27,7 +27,7 @@ class MenuGroupServiceTest {
         MenuGroupCreateRequest request = new MenuGroupCreateRequest("단일 메뉴");
         MenuGroupResponse response = menuGroupService.create(request);
 
-        MenuGroup findMenuGroup = menuGroupDao.findById(response.getId())
+        MenuGroup findMenuGroup = menuGroupRepository.findById(response.getId())
                 .orElseThrow(RuntimeException::new);
 
         assertThat(findMenuGroup.getName()).isEqualTo(response.getName());
@@ -39,14 +39,14 @@ class MenuGroupServiceTest {
         MenuGroup menuGroup1 = new MenuGroup("단일 메뉴");
         MenuGroup menuGroup2 = new MenuGroup("세트 메뉴");
 
-        menuGroupDao.save(menuGroup1);
-        menuGroupDao.save(menuGroup2);
+        menuGroupRepository.save(menuGroup1);
+        menuGroupRepository.save(menuGroup2);
 
         assertThat(menuGroupService.list()).hasSize(2);
     }
 
     @AfterEach
     void tearDown() {
-        menuGroupDao.deleteAll();
+        menuGroupRepository.deleteAll();
     }
 }
